requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2025.8.3
click==8.3.0
dnspython==2.8.0
Flask==3.1.2
Flask-Session==0.8.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.19.0
packaging==25.0
pymongo==4.15.1
python-dotenv==1.1.1
Werkzeug==3.1.3
